Revert "Modify loadBrain to use imports instead of dynamic path loading (#865)"
This reverts commit 6d916bda7d8ac3397220c3784a2eb4c06b0311ef.
