+++ src/utils/loadBrain.ts
@@ -1,55 +1,49 @@
-import { apis } from '@/brain/apis';
-import { appHome } from '@/brain/appHome';
-import { ghaCancel } from '@/brain/ghaCancel';
-import { githubMetrics } from '@/brain/githubMetrics';
-import { gocdConsecutiveUnsuccessfulAlert } from '@/brain/gocdConsecutiveUnsuccessfulAlert';
-import { gocdDataDog } from '@/brain/gocdDataDog';
-import { gocdNoDeploysAlert } from '@/brain/gocdNoDeploysAlert';
-import { gocdSlackFeeds } from '@/brain/gocdSlackFeeds';
-import { issueLabelHandler } from '@/brain/issueLabelHandler';
-import { issueNotifier } from '@/brain/issueNotifier';
-import { notificationPreferences } from '@/brain/notificationPreferences';
-import { notifyOnGoCDStageEvent } from '@/brain/notifyOnGoCDStageEvent';
-import { pleaseDeployNotifier } from '@/brain/pleaseDeployNotifier';
-import { projectsHandler } from '@/brain/projectsHandler';
-import { requiredChecks } from '@/brain/requiredChecks';
-import { saveGoCDStageEvents } from '@/brain/saveGoCDStageEvents';
-import { syncSlackUsers } from '@/brain/syncSlackUsers';
-import { syncUserProfileChange } from '@/brain/syncUserProfileChange';
-import { triggerPubSub } from '@/brain/triggerPubSub';
-import { typescript } from '@/brain/typescript';
+import { promises as fs } from 'fs';
+import path from 'path';
+
+import * as Sentry from '@sentry/node';
+
+const ROOT = path.join(__dirname, '../brain');
 
 /**
- * Loads all functions in `@/brain`
+ * Loads all modules in the root of `@/brain`
  *
  * Currently we only want to load the exported fn whose name matches the module name,
  * this is because we sometimes export functions only for testing.
  */
-export async function loadBrain(): Promise<string[]> {
-  const loadFunctions = [
-    apis,
-    appHome,
-    ghaCancel,
-    githubMetrics,
-    gocdConsecutiveUnsuccessfulAlert,
-    gocdDataDog,
-    gocdNoDeploysAlert,
-    gocdSlackFeeds,
-    issueLabelHandler,
-    issueNotifier,
-    notificationPreferences,
-    notifyOnGoCDStageEvent,
-    pleaseDeployNotifier,
-    projectsHandler,
-    requiredChecks,
-    saveGoCDStageEvents,
-    syncSlackUsers,
-    syncUserProfileChange,
-    triggerPubSub,
-    typescript,
-  ];
-  loadFunctions.forEach((f) => f());
+export async function loadBrain() {
+  const modules = getExportedFunctions(await getBrainModules());
+  modules.forEach((m) => m());
+}
 
-  // For test purposes, return the names of the functions loaded
-  return loadFunctions.map((f) => f.name);
+/**
+ * This is only exported for test
+ */
+export async function getBrainModules() {
+  return (await fs.readdir(ROOT)).filter(
+    (f) => !f.endsWith('.d.ts') && !f.endsWith('.map') && !f.endsWith('.md')
+  );
+}
+
+/**
+ * This is only exported for test
+ */
+export function getExportedFunctions(fileNames: string[]) {
+  return fileNames.flatMap((f) => {
+    try {
+      // Only return imported functions that match filename
+      // This is because we sometimes need to export other functions to test
+      return (
+        Object.entries(require(path.join(ROOT, f)))
+          // @ts-ignore
+          .filter(([key]) => key === f)
+          .map(([, value]) => value)
+      );
+    } catch (e) {
+      const err = new Error(`Unable to load brain: ${f}`);
+      Sentry.captureException(err);
+      console.error(err);
+      return [];
+    }
+  }) as Function[];
 }
